Keep trainable parameters in float32 under mixed precision (#5)
model_dtype previously cast the whole module tree, LoRA adapters
included, to bf16. At the configured lr of 2.5e-5 every Adam update to
the adapters rounded to zero against bf16 resolution, so autoregressive
OpenVLA never learned the action-token mapping: 2/130 on libero_spatial
with lora_A still at its exact init after 84k optimizer steps.

- _apply_model_dtype() re-casts requires_grad parameters to float32 for
  the mixed precisions; frozen parameters and buffers keep the
  low-precision dtype. PrecisionType gains is_mixed() and autocast().
- set_image_size mock forwards run under autocast so shape inference
  works on mixed float32/low-precision encoders.
- SyntheticRolloutCallback, PriorTargetStandardizationCallback, and
  explainability attribution run their forwards under the training
  autocast; previously they crashed for any bf16-mixed policy.
- DFormerV2.set_image_size re-freezes the rebuilt pooling head on
  frozen encoders.
- The heatmap writer casts to float32 before numpy conversion.
- OpenVLA and OpenVLA-OFT presets train LoRA at the recipe lr of 5e-4.

Verified: full unit and integration suites pass, and a one-epoch
bf16-mixed synthetic training runs end to end (it crashes on main).

+++ src/versatil/explainability/writer.py
@@ -68,7 +68,7 @@ class ExplanationWriter:
             {
                 "metadata": metadata,
                 "heatmaps": {
-                    camera: heatmap.detach().cpu()
+                    camera: heatmap.detach().float().cpu()
                     for camera, heatmap in heatmaps.items()
                 },
             },
@@ -103,7 +103,7 @@ class ExplanationWriter:
                 raise RuntimeError(
                     f"No display observation found for heatmap camera '{camera}'."
                 )
-            cpu_heatmap = heatmap.detach().cpu()
+            cpu_heatmap = heatmap.detach().float().cpu()
             for batch_index in range(cpu_heatmap.shape[0]):
                 sample_label = self.sample_label(
                     metadata=batch.metadata,
